# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GTD Backend API (Collapsed)",
    "version": "1.0.3",
    "description": "Fully collapsed OpenAPI spec for Todoist via the GTD backend (≤30 operations)"
  },
  "servers": [
    {
      "url": "https://gtd-backend.onrender.com"
    }
  ],
  "paths": {
    "/tasks/manage": {
  "post": {
    "operationId": "manageTasks",
    "summary": "List, get, create, update, delete, move or change status of tasks",
    "requestBody": {
      "required": true,
      "content": {
        "application/json": {
          "schema": {
            "type": "object",
            "properties": {
              "action": {
                "type": "string",
                "enum": ["list", "get", "create", "update", "delete", "move", "status"]
              },
              "task_id": {
                "type": "string",
                "description": "Required for get, update, delete, move, status actions."
              },
              "project_id": {
                "type": "string",
                "description": "Optional for create/list/move. Used to specify the project. Use filters for more specific listing."
              },
              "section_id": {
                "type": "string",
                "description": "Optional for create/move. Used to specify the section within a project."
              },
              "content": {
                "type": "string",
                "description": "Required for create, optional for update. The task's content."
              },
              "description": {
                "type": "string",
                "description": "Optional for create/update. Supplementary details for the task."
              },
              "due_string": {
                "type": "string",
                "description": "Optional for create/update. Human-readable due date (e.g., 'tomorrow', 'next monday')."
              },
              "priority": {
                "type": "integer",
                "minimum": 1,
                "maximum": 4,
                "description": "Optional for create/update. Task priority (1=p4, 4=p1)."
              },
              "labels": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "Optional for create/update. Array of label names (strings) to attach (e.g., [\"email\", \"work\"])"
              },
              "status": {
                "type": "string",
                "enum": ["closed", "open"],
                "description": "Required for status action. Use 'closed' to complete, 'open' to reopen."
              },
              "filter": {
                "type": "string",
                "description": "Optional for 'list' action. Todoist filter query string (e.g., 'today', 'p1 & #Work', '@waiting_for'). Helps avoid ResponseTooLarge errors."
              },
              "label_id": {
                "type": "string",
                "description": "Optional for 'list' action. Filter tasks by a specific label ID."
              }
            },
            "required": ["action"]
          }
        }
      }
    },
    "responses": {
      "200": {
        "description": "JSON response for list/get/create/update/move actions."
      },
      "204": {
        "description": "No content response for successful delete/status actions."
      },
      "400": {
        "description": "Bad Request - Invalid input, missing required fields, or filter needed for large lists."
      },
      "404": {
        "description": "Not Found - Task, project, or section ID does not exist."
      },
      "413": {
        "description": "Payload Too Large / Response Too Large - Consider using filters for 'list' action."
      }
    }
  }
},
    "/projects/manage": {
      "post": {
        "operationId": "manageProjects",
        "summary": "List, get, create, update, delete, or list collaborators of projects",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": [
                      "list",
                      "get",
                      "create",
                      "update",
                      "delete",
                      "collaborators"
                    ]
                  },
                  "project_id": {
                    "type": "string",
                    "description": "Required for get, update, delete, collaborators actions."
                  },
                  "name": {
                    "type": "string",
                    "description": "Required for create, optional for update."
                  },
                  "color": {
                    "type": "string",
                    "description": "Optional for create/update."
                  },
                  "parent_id": {
                    "type": "string",
                    "description": "Optional for create."
                  },
                  "is_favorite": {
                    "type": "boolean",
                    "description": "Optional for create/update."
                  }
                },
                "required": [
                  "action"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JSON response for list/get/create/update/collaborators actions."
          },
          "204": {
            "description": "No content response for successful delete action."
          }
        }
      }
    },
    "/sections/manage": {
      "post": {
        "operationId": "manageSections",
        "summary": "List, get, create, update or delete sections",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": [
                      "list",
                      "get",
                      "create",
                      "update",
                      "delete"
                    ]
                  },
                  "section_id": {
                    "type": "string",
                     "description": "Required for get, update, delete actions."
                  },
                  "project_id": {
                    "type": "string",
                    "description": "Required for create, optional for list actions."
                  },
                  "name": {
                    "type": "string",
                     "description": "Required for create, optional for update."
                  },
                  "order": {
                    "type": "integer",
                     "description": "Optional for create/update."
                  }
                },
                "required": [
                  "action"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JSON response for list/get/create/update actions."
          },
          "204": {
            "description": "No content response for successful delete action."
          }
        }
      }
    },
    "/labels/manage": {
      "post": {
        "operationId": "manageLabels",
        "summary": "List, get, create, update, delete personal labels",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": [
                      "list",
                      "get",
                      "create",
                      "update",
                      "delete"
                    ]
                  },
                  "label_id": {
                    "type": "string",
                     "description": "Required for get, update, delete actions (usually a string ID)."
                  },
                  "name": {
                    "type": "string",
                     "description": "Required for create, optional for update."
                  },
                  "color": {
                    "type": "string",
                     "description": "Optional for create/update."
                  },
                  "order": {
                    "type": "integer",
                     "description": "Optional for create/update."
                  },
                  "is_favorite": {
                    "type": "boolean",
                     "description": "Optional for create/update."
                  }
                },
                "required": [
                  "action"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JSON response for list/get/create/update actions."
          },
          "204": {
            "description": "No content response for successful delete action."
          }
        }
      }
    },
    "/comments/manage": {
      "post": {
        "operationId": "manageComments",
        "summary": "List, get, create, update or delete comments",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": [
                      "list",
                      "get",
                      "create",
                      "update",
                      "delete"
                    ]
                  },
                  "comment_id": {
                    "type": "string",
                     "description": "Required for get, update, delete actions."
                  },
                  "task_id": {
                    "type": "string",
                     "description": "Required for create, required for list if project_id not provided."
                  },
                  "project_id": {
                    "type": "string",
                     "description": "Optional for create, required for list if task_id not provided."
                  },
                  "content": {
                    "type": "string",
                     "description": "Required for create/update."
                  }
                },
                "required": [
                  "action"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JSON response for list/get/create/update actions."
          },
          "204": {
            "description": "No content response for successful delete action."
          }
        }
      }
    },
    "/collaborators/manage": {
      "post": {
        "operationId": "manageCollaborators",
        "summary": "List collaborators on shared projects",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": [
                      "list"
                    ],
                     "description": "Only 'list' is supported."
                  },
                  "project_id": {
                    "type": "string",
                     "description": "Required for list action."
                  }
                },
                "required": [
                  "action",
                  "project_id"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JSON response for list collaborators."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {},
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY"
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ]
}